requirements.txt
fastapi>=0.111
uvicorn>=0.30
requests>=2.31
pydantic>=2.7
websocket-client>=1.7
PyYAML>=6.0
openai>=1.40
wyoming>=1.5
audioop-lts; python_version >= "3.13"
